+++ src/App.tsx
@@ -2,14 +2,12 @@ import * as React from 'react';
 import { ChakraProvider } from '@chakra-ui/react';
 import theme from '@style/theme';
 import Router from '@pages/Router';
-import Modal from '@components/Modal';
 import { RecoilRoot } from 'recoil';
 
 export const App = () => (
   <RecoilRoot>
     <ChakraProvider theme={theme}>
       <Router />
-      <Modal />
     </ChakraProvider>
   </RecoilRoot>
 );
