+++ src/components/AlertButton.tsx
@@ -1,0 +1,68 @@
+import {
+  AlertDialog,
+  AlertDialogBody,
+  AlertDialogContent,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogOverlay,
+  Button,
+  useDisclosure,
+} from '@chakra-ui/react';
+import React from 'react';
+
+interface Props {
+  color?: string;
+  btnName: string;
+  alertTitle: string;
+  alertText: string;
+  callback: () => void;
+}
+
+const AlertButton: React.FC<Props> = ({
+  color = 'red',
+  btnName,
+  alertTitle,
+  alertText,
+  callback,
+}) => {
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const cancelRef = React.useRef(null);
+
+  return (
+    <>
+      <Button colorScheme={color} onClick={onOpen}>
+        {btnName}
+      </Button>
+
+      <AlertDialog isOpen={isOpen} leastDestructiveRef={cancelRef} onClose={onClose}>
+        <AlertDialogOverlay>
+          <AlertDialogContent>
+            <AlertDialogHeader fontSize="lg" fontWeight="bold">
+              {alertTitle}
+            </AlertDialogHeader>
+
+            <AlertDialogBody>{alertText}</AlertDialogBody>
+
+            <AlertDialogFooter>
+              <Button ref={cancelRef} onClick={onClose}>
+                Cancel
+              </Button>
+              <Button
+                colorScheme="green"
+                onClick={() => {
+                  onClose();
+                  callback();
+                }}
+                ml={3}
+              >
+                Confirm
+              </Button>
+            </AlertDialogFooter>
+          </AlertDialogContent>
+        </AlertDialogOverlay>
+      </AlertDialog>
+    </>
+  );
+};
+
+export default AlertButton;
